Code review based on OCA guidelines
