+++ business_requirement_deliverable/models/__init__.py
@@ -1,4 +1,7 @@
 # -*- coding: utf-8 -*-
 # © 2016 Elico Corp (https://www.elico-corp.com).
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-from . import business
+
+from . import business_requirement
+from . import business_requirement_deliverable
+from . import business_requirement_resource
