+++ business_requirement_deliverable/security/ir.model.access.csv
@@ -1,6 +1,5 @@
 id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
 access_business_requirement_deliverable_user,business.requirement.deliverable.user,model_business_requirement_deliverable,business_requirement.group_business_requirement_user,1,1,1,1
 access_business_requirement_deliverable_manager,business.requirement.deliverable.manager,model_business_requirement_deliverable,business_requirement.group_business_requirement_manager,1,1,1,1
-
 access_business_requirement_resource_user,business.requirement.resource.user,model_business_requirement_resource,business_requirement.group_business_requirement_user,1,1,1,1
 access_business_requirement_resource_manager,business.requirement.resource.manager,model_business_requirement_resource,business_requirement.group_business_requirement_manager,1,1,1,1
